+++ packages/react-query/src/__tests__/HydrationBoundary.test.tsx
@@ -211,13 +211,10 @@ describe('React hydration', () => {
       const newDehydratedState = dehydrate(intermediateClient)
       intermediateClient.clear()
 
-      function Thrower() {
+      function Thrower(): never {
         throw new Promise(() => {
           // Never resolve
         })
-
-        // @ts-expect-error
-        return null
       }
 
       React.startTransition(() => {
@@ -482,5 +479,65 @@ describe('React hydration', () => {
     hydrateSpy.mockRestore()
     prefetchQueryClient.clear()
     clientQueryClient.clear()
+  })
+
+  test('should not refetch when query has enabled set to false', async () => {
+    const queryFn = vi.fn()
+    const queryClient = new QueryClient()
+
+    function Page() {
+      const { data } = useQuery({
+        queryKey: ['string'],
+        queryFn,
+        enabled: false,
+      })
+      return <div>{JSON.stringify(data)}</div>
+    }
+
+    const rendered = render(
+      <QueryClientProvider client={queryClient}>
+        <HydrationBoundary state={JSON.parse(stringifiedState)}>
+          <Page />
+        </HydrationBoundary>
+      </QueryClientProvider>,
+    )
+
+    expect(rendered.getByText('["stringCached"]')).toBeInTheDocument()
+
+    await vi.advanceTimersByTimeAsync(11)
+    expect(queryFn).toHaveBeenCalledTimes(0)
+    expect(rendered.getByText('["stringCached"]')).toBeInTheDocument()
+
+    queryClient.clear()
+  })
+
+  test('should not refetch when query has staleTime set to Infinity', async () => {
+    const queryFn = vi.fn()
+    const queryClient = new QueryClient()
+
+    function Page() {
+      const { data } = useQuery({
+        queryKey: ['string'],
+        queryFn,
+        staleTime: Infinity,
+      })
+      return <div>{JSON.stringify(data)}</div>
+    }
+
+    const rendered = render(
+      <QueryClientProvider client={queryClient}>
+        <HydrationBoundary state={JSON.parse(stringifiedState)}>
+          <Page />
+        </HydrationBoundary>
+      </QueryClientProvider>,
+    )
+
+    expect(rendered.getByText('["stringCached"]')).toBeInTheDocument()
+
+    await vi.advanceTimersByTimeAsync(11)
+    expect(queryFn).toHaveBeenCalledTimes(0)
+    expect(rendered.getByText('["stringCached"]')).toBeInTheDocument()
+
+    queryClient.clear()
   })
 })