+++ packages/vue-query-devtools/CHANGELOG.md
@@ -1,5 +1,45 @@
 # @tanstack/vue-query-devtools
 
+## 6.1.3
+
+### Patch Changes
+
+- Updated dependencies [[`f9fc56a`](https://github.com/TanStack/query/commit/f9fc56a9b8724bcfae46f8f6cb229123478eb4db), [`0b29b6f`](https://github.com/TanStack/query/commit/0b29b6f877d4b3a6d05b1c85fb9cb1e6ea736291)]:
+  - @tanstack/query-devtools@5.92.0
+  - @tanstack/vue-query@5.92.3
+
+## 6.1.2
+
+### Patch Changes
+
+- Updated dependencies [[`758414f`](https://github.com/TanStack/query/commit/758414fcecb9b56142014edf3ff455c283fb556e)]:
+  - @tanstack/vue-query@5.92.0
+
+## 6.1.1
+
+### Patch Changes
+
+- Updated dependencies [[`b261b6f`](https://github.com/TanStack/query/commit/b261b6f29eee2a9bdbe1bc20035fe9b83b15376b)]:
+  - @tanstack/query-devtools@5.91.1
+
+## 6.1.0
+
+### Minor Changes
+
+- feat(devtools): allow passing a theme via prop ([#9887](https://github.com/TanStack/query/pull/9887))
+
+### Patch Changes
+
+- Updated dependencies [[`0e9d5b5`](https://github.com/TanStack/query/commit/0e9d5b565276f0de2a1a14ffbb079b5988581c27)]:
+  - @tanstack/query-devtools@5.91.0
+
+## 6.0.0
+
+### Patch Changes
+
+- Updated dependencies [[`b2bd79d`](https://github.com/TanStack/query/commit/b2bd79d0a6b2707461897c426b0d2275a3318e4b)]:
+  - @tanstack/vue-query@5.91.0
+
 ## 5.91.0
 
 ### Minor Changes